Refactored task results from task stages file, endpoint completion
In order to keep things organized and clean, I've moved the task result API definitions from
the task_stages.go file. I've also added the Read endpoint for task results, although I'm hesistant
to include this in the release as individual task results are rarely fetched since they are
included when you poll a task stage (more discovery would be needed).
