+++ task_stages.go
@@ -85,33 +85,3 @@ func (s *taskStages) List(ctx context.Context, runID string, options *TaskStageL
 
 	return tlist, nil
 }
-
-type TaskResultStatus string
-type TaskEnforcementLevel string
-
-const (
-	TaskPassed      TaskResultStatus     = "passed"
-	TaskFailed      TaskResultStatus     = "failed"
-	TaskRunning     TaskResultStatus     = "running"
-	TaskPending     TaskResultStatus     = "pending"
-	TaskUnreachable TaskResultStatus     = "unreachable"
-	Advisory        TaskEnforcementLevel = "advisory"
-	Mandatory       TaskEnforcementLevel = "mandatory"
-)
-
-type TaskResult struct {
-	ID                            string                  `jsonapi:"primary,task-results"`
-	Status                        TaskResultStatus        `jsonapi:"attr,status"`
-	Message                       string                  `jsonapi:"attr,message"`
-	StatusTimestamps              RunTaskStatusTimestamps `jsonapi:"attr,status-timestamps"`
-	URL                           string                  `jsonapi:"attr,url"`
-	CreatedAt                     time.Time               `jsonapi:"attr,created-at,iso8601"`
-	UpdatedAt                     time.Time               `jsonapi:"attr,updated-at,iso8601"`
-	TaskID                        string                  `jsonapi:"attr,task-id"`
-	TaskName                      string                  `jsonapi:"attr,task-name"`
-	TaskURL                       string                  `jsonapi:"attr,task-url"`
-	WorkspaceTaskID               string                  `jsonapi:"attr,workspace-task-id"`
-	WorkspaceTaskEnforcementLevel TaskEnforcementLevel    `jsonapi:"attr,workspace-task-enforcement-level"`
-
-	TaskStage *TaskStage `jsonapi:"relation,task_stage"`
-}
